% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpointjoin.R
\name{endpointjoin}
\alias{endpointjoin}
\title{endpointjoin  
For fʌn.nɔlɛdʒi (KS Dissertation)  
Combine Looking-While-Listening Data with PsychoPy Endpoint Touchscreen Data  
  
This takes LWL data from DataWiz that has been compiled and processed in excel and combines it with PsychoPy Data.    
The two .CSV file inputs should have the same number of rows. 
  
LWLcsv Data should have:   
A WIDE format (each timepoint is a column and each row is a trial where there are 38 trials per subject)   
Backfilled rows  
Columns up to F2400  
GoodLooks column is added to the dataset  
  
PsychoPycsv Data should:  
Be compiled across ALL subjects (using a postprocess_LWL function)  
  
N.B. Example in this documentation will only work in the LWL.Rproj}
\usage{
endpointjoin(LWLcsv, PsychoPycsv, experiment, group)
}
\arguments{
\item{LWLcsv}{.CSV filename (string) with cleaned LWL data output from DataWiz (wide format)}

\item{PsychoPycsv}{.CSV filename (string) with all PsychoPy trials - this should have the same number of rows as the LWLcsv input}

\item{experiment}{either experiment 1 (NO prime) or experiment 2 (with semantic prime). Must be a numeric value.}

\item{group}{String value for Hearing Group (either "NH" for normal hearing or "HL" for hearing loss). This will be inputted into filename}
}
\value{
Saves a file with LWL and endpoint accuracy data combined in a single CSV. Saves the dataframe in environment as "FUNK_combined"
}
\description{
endpointjoin  
For fʌn.nɔlɛdʒi (KS Dissertation)  
Combine Looking-While-Listening Data with PsychoPy Endpoint Touchscreen Data  
  
This takes LWL data from DataWiz that has been compiled and processed in excel and combines it with PsychoPy Data.    
The two .CSV file inputs should have the same number of rows. 
  
LWLcsv Data should have:   
A WIDE format (each timepoint is a column and each row is a trial where there are 38 trials per subject)   
Backfilled rows  
Columns up to F2400  
GoodLooks column is added to the dataset  
  
PsychoPycsv Data should:  
Be compiled across ALL subjects (using a postprocess_LWL function)  
  
N.B. Example in this documentation will only work in the LWL.Rproj
}
\examples{
endpointjoin("01data/Merging/2019-09-23-E1_LWL_HL_Compiled_forMerge.csv",
"01data/Merging/2019-09-09-Exp1_HL_Endpoint_forMerge.csv", 1,"HL") 
# Excuse the poor filenaming ... eek!


}
