% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longLWL.R
\name{longLWL}
\alias{longLWL}
\title{longLWL
For fʌn.nɔlɛdʒi (KS Dissertation)
Make Looking-While-Listening Data from DataWiz into long form}
\usage{
longLWL(filename, experiment)
}
\arguments{
\item{filename}{a string that indicates the filename for the LWL data (including path)}

\item{experiment}{either experiment 1 (NO prime) or experiment 2 (with semantic prime). Must be a numeric value.}
}
\value{
a data frame labelled LWL_long in your environment.
}
\description{
This function will take LWL data in its CLEANED form and gathers data so Timebin and Fixations are two columns instead of each timebin being a column with fixation values for each participant.
LWL data is CLEANED if:
A GoodLooks column is added to the dataset
The LWL trials are paired with Endpoint accuracy from PsychoPy
}
\details{
This function will turn all center looks (0.5 fixation values) into "C" non-numeric indications for analysis.
This function will also get rid of file extensions for image and audio filenames.
}
